requirements-prod.txt
# FastAPI Backend Requirements for Production Installation

# Core Web Framework
fastapi>=0.109.0,<0.115.0
uvicorn[standard]>=0.27.0,<0.31.0
websockets>=12.0,<13.0
strawberry-graphql>=0.218.0

# Async Database
psycopg2-binary>=2.9.0,<3.0
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.13.0,<2.0.0
asyncpg>=0.28.0,<0.31.0

# Authentication & Security
pydantic>=2.5.0,<2.9.0
pydantic-settings>=2.1.0,<2.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.6,<0.1.0
passlib[bcrypt]>=1.7.4,<1.8.0
authlib>=1.3.0,<1.4.0

# Async HTTP Client
httpx>=0.26.0,<0.28.0

# Testing (optional for production)
pytest>=7.4.0,<7.5.0
pytest-asyncio>=0.24.0,<0.25.0

# Code Quality (optional)
black>=24.0.0,<25.0.0
ruff>=0.1.0,<0.5.0
mypy>=1.8.0,<1.10.0
